Map , deltatime Fixed,
change datatype float for delta time to double

+++ Enemy.cpp
@@ -18,7 +18,7 @@ void Enemy::initialize()
 	boundingRectangle.setSize(sf::Vector2f(size.x * Sprite.getScale().x, size.y * Sprite.getScale().y));
 }
 
-void Enemy::Update(float deltaTime)
+void Enemy::Update(double deltaTime)
 {
 
 	if (health > 0) {
